Oh yeah edge stereotypes are added
