+++ src/himeko/hbcm/elements/vertex.py
@@ -5,7 +5,6 @@ from himeko.hbcm.elements.element import HypergraphElement
 from himeko.hbcm.exceptions.basic_exceptions import (InvalidHypergraphElementException,
                                                      InvalidParentException, ElementSelfParentException)
 
-from collections import deque
 
 
 class HyperVertex(HypergraphElement):
@@ -26,10 +25,6 @@ class HyperVertex(HypergraphElement):
         # Add current element to parent
         if parent is not None:
             parent.add_element(self)
-        self._elements: typing.Dict[bytes, HypergraphElement] = {}
-
-        # Indexing
-        self.__index_named_elements: typing.Dict[str, HypergraphElement] = {}
         # Adding logger element (ya rly)
         self._logger = logging.getLogger(f"{self.label}")
 
@@ -48,7 +43,7 @@ class HyperVertex(HypergraphElement):
         if not isinstance(v, HypergraphElement):
             raise InvalidHypergraphElementException("Unable to add incompatible element")
         self._elements[v.guid] = v
-        self.__index_named_elements[v.name] = v
+        self._index_named_elements[v.name] = v
         # Set element parent (if parent is not already self)
         if v.parent is not self:
             v._parent = self
@@ -62,14 +57,14 @@ class HyperVertex(HypergraphElement):
         self._elements.pop(v.guid)
 
     def update_element(self, v: HypergraphElement):
-        if v.name not in self.__index_named_elements:
+        if v.name not in self._index_named_elements:
             self.add_element(v)
         else:
             # Remove existing element
-            __tmp = self.__index_named_elements[v.name]
+            __tmp = self._index_named_elements[v.name]
             self._elements.pop(__tmp.guid)
-            self.__index_named_elements[v.name] = v
-            self.__index_named_elements[v.guid] = v
+            self._index_named_elements[v.name] = v
+            self._index_named_elements[v.guid] = v
 
     def __iadd__(self, other):
         if isinstance(other, typing.Iterable):
@@ -102,41 +97,11 @@ class HyperVertex(HypergraphElement):
     def __hash__(self):
         return int.from_bytes(self.guid, "big")
 
-    def __fringe_append(self, fringe, __e, d):
-        for _, ch in __e._elements.items():
-            fringe.appendleft((ch, d + 1))
 
-    def __init_fringe(self, include_self: bool):
-        visited = set()
-        fringe = deque()
-        if include_self:
-            fringe.append((self, 0))
-        else:
-            self.__fringe_append(fringe, self, 0)
-        return visited, fringe
 
-    def get_subelements(self,
-                        condition: typing.Callable[[HypergraphElement], bool],
-                        depth: typing.Optional[int] = None, include_self=False):
-        visited, fringe = self.__init_fringe(include_self)
-        __res = []
-        if depth is None:
-            bound_condition = lambda x: True
-        else:
-            bound_condition = lambda x: d <= depth
-        while len(fringe) != 0:
-            __e, d = fringe.pop()
-            if __e not in visited and bound_condition(d):
-                visited.add(__e)
-                if condition(__e):
-                    __res.append(__e)
-                    yield __e
-                if isinstance(__e, HyperVertex):
-                    self.__fringe_append(fringe, __e, d)
-        return __res
 
-    def get_children(self, condition: typing.Callable[[HypergraphElement], bool], depth: typing.Optional[int] = 1):
-        return self.get_subelements(condition, depth)
+
+
 
     def get_children_nodes(self, condition: typing.Callable[[HypergraphElement], bool], depth: typing.Optional[int] = 1):
         return self.get_subelements(lambda x:
